refactor(settings): extract Telemetry section to _components/
PR-2 of the settings-page split. Pulls the entire Telemetry
section (#s-telemetry) into a self-contained TelemetrySection
component.

Moved into the child: 9 useState/useRef hooks, both
telemetry-prefs fetch effects, saveTelemetryPref, and
resetTelemetryData. The section talks only to
/api/bridge/telemetry-prefs — no shared `settings` slice — so the
only parent dependency is `flashSaved` (the global "Saved" toast),
passed as a prop.

The optimistic-save AbortController, previously the page-level
`abortRef`, is now owned by the section and aborts on its own
unmount.

Behavior identical. page.tsx: 1971 → 1710 lines. A keystroke
elsewhere on the settings page no longer re-runs the Telemetry
subtree.
